build/board: refactor and reorganize sources/inclusions/paths
The reasoning behind most of these changes is sharing of core module
sources in different projects. Therefore, sources have been split up and
made more generic where possible, unneeded includes have been removed etc.

+++ src/application/database/Database.h
@@ -171,7 +171,7 @@ namespace Database
         uint32_t _lastPresetAddress;
 
         /// Holds total number of supported presets.
-        uint16_t _supportedPresets;
+        size_t _supportedPresets;
 
         uint16_t _uid = 0;
